+++ R/marginal-effects.R
@@ -16,7 +16,7 @@
 #' \donttest{## Generate synthetic data.
 #' set.seed(1986)
 #' 
-#' data <- generate_ordered_data(1000)
+#' data <- generate_ordered_data(100)
 #' sample <- data$sample
 #' Y <- sample$Y
 #' X <- sample[, -1]