+++ NEWS.md
@@ -1,3 +1,11 @@
+## version 1.0.2
+
+Fixed an issue arising with check on outcome formatting. 
+
+Fixed an issue arising when one class of the outcome has zero observations.
+
+Updated references.
+
 ## version 1.0.1
 
 Now the ocf function checks whether the outcome is correctly coded.